fix: disable buy/sell when token needs graduation
Move Graduate button from TokenDetail to BuySellPanel.
When bonding curve is 100% complete but not graduated yet,
hide buy/sell inputs and show graduation prompt instead.
