+++ apps/web/src/pages/Launchpad/components/BuySellPanel.tsx
@@ -4,7 +4,7 @@ import { useAccountDrawer } from 'components/AccountDrawer/MiniPortfolio/hooks'
 import { ToastRegularSimple } from 'components/Popups/ToastRegularSimple'
 import { useAccount } from 'hooks/useAccount'
 import { useBondingCurveBalance, useCalculateBuy, useCalculateSell } from 'hooks/useBondingCurveToken'
-import { calculateMinOutput, useBuy, useSell } from 'hooks/useLaunchpadActions'
+import { calculateMinOutput, useBuy, useGraduate, useSell } from 'hooks/useLaunchpadActions'
 import { useTokenAllowance, useUpdateTokenAllowance } from 'hooks/useTokenAllowance'
 import styledComponents from 'lib/styled-components'
 import { useCallback, useMemo, useState } from 'react'
@@ -165,12 +165,26 @@ const TradeOnSwapButton = styled(Flex, {
   },
 })
 
+const GraduateButton = styled(Flex, {
+  alignItems: 'center',
+  justifyContent: 'center',
+  paddingVertical: '$spacing16',
+  backgroundColor: '$accent1',
+  borderRadius: '$rounded12',
+  cursor: 'pointer',
+  hoverStyle: {
+    backgroundColor: '$accent2',
+  },
+})
+
 interface BuySellPanelProps {
   tokenAddress: string
   tokenSymbol: string
   baseAsset: string
   graduated: boolean
+  canGraduate: boolean
   onTransactionComplete?: () => void
+  onGraduateComplete?: () => void
 }
 
 export function BuySellPanel({
@@ -178,11 +192,14 @@ export function BuySellPanel({
   tokenSymbol,
   baseAsset,
   graduated,
+  canGraduate,
   onTransactionComplete,
+  onGraduateComplete,
 }: BuySellPanelProps) {
   const [isBuy, setIsBuy] = useState(true)
   const [amount, setAmount] = useState('')
   const [isLoading, setIsLoading] = useState(false)
+  const [isGraduating, setIsGraduating] = useState(false)
   const [error, setError] = useState<string | null>(null)
 
   const navigate = useNavigate()
@@ -316,6 +333,7 @@ export function BuySellPanel({
   // Action hooks
   const buy = useBuy(tokenAddress, chainId)
   const sell = useSell(tokenAddress, chainId)
+  const graduate = useGraduate(tokenAddress, chainId)
 
   // Format output
   const outputAmount = useMemo(() => {
@@ -358,6 +376,41 @@ export function BuySellPanel({
       setAmount(formatUnits(tokenBalance, 18))
     }
   }, [isBuy, baseBalance, tokenBalance])
+
+  const handleGraduate = useCallback(async () => {
+    setIsGraduating(true)
+    try {
+      const tx = await graduate()
+      addTransaction(tx, {
+        type: TransactionType.LaunchpadGraduate,
+        tokenAddress: assume0xAddress(tokenAddress),
+        dappInfo: { name: `Graduated ${tokenSymbol} to JuiceSwap V2` },
+      })
+      await tx.wait()
+      toast(
+        <ToastRegularSimple
+          icon={<CheckCircleFilled color="$statusSuccess" size="$icon.28" />}
+          text={
+            <Flex gap="$gap4" flexWrap="wrap" flex={1}>
+              <Text variant="body2" color="$neutral1">
+                Graduated
+              </Text>
+              <Text variant="body3" color="$neutral2">
+                {tokenSymbol} to JuiceSwap V2
+              </Text>
+            </Flex>
+          }
+          onDismiss={() => toast.dismiss()}
+        />,
+        { duration: 5000 },
+      )
+      onGraduateComplete?.()
+    } catch {
+      setError('Failed to graduate token')
+    } finally {
+      setIsGraduating(false)
+    }
+  }, [graduate, addTransaction, tokenAddress, tokenSymbol, onGraduateComplete])
 
   const handleAction = useCallback(async () => {
     if (!parsedAmount || parsedAmount === 0n || !account.address) {
@@ -486,9 +539,14 @@ export function BuySellPanel({
     queryClient,
   ])
 
+  const needsGraduation = canGraduate && !graduated
+
   const buttonText = useMemo(() => {
     if (!account.address) {
       return 'Connect Wallet'
+    }
+    if (needsGraduation) {
+      return 'Graduate first'
     }
     if (isLoading) {
       return 'Processing...'
@@ -515,11 +573,13 @@ export function BuySellPanel({
     needsBaseApproval,
     needsTokenApproval,
     tokenSymbol,
+    needsGraduation,
   ])
 
   const isWalletConnected = !!account.address
   const isButtonDisabled =
-    isWalletConnected && (isLoading || !parsedAmount || parsedAmount === 0n || hasInsufficientBalance)
+    isWalletConnected &&
+    (isLoading || !parsedAmount || parsedAmount === 0n || hasInsufficientBalance || needsGraduation)
 
   const handleButtonPress = useCallback(() => {
     if (!isWalletConnected) {
@@ -562,75 +622,94 @@ export function BuySellPanel({
         </Tab>
       </TabContainer>
 
-      <InputContainer>
-        <InputLabel>{isBuy ? 'You pay (JUSD)' : `You sell (${tokenSymbol})`}</InputLabel>
-        <InputWrapper>
-          <StyledInput
-            type="text"
-            inputMode="decimal"
-            placeholder="0.0"
-            value={amount}
-            onChange={handleAmountChange}
-            autoComplete="off"
-            autoCorrect="off"
-          />
-          <MaxButton onPress={handleMaxClick}>
-            <Text variant="buttonLabel4" color="$accent1">
-              MAX
+      {needsGraduation ? (
+        <>
+          <Text variant="body2" color="$neutral2" textAlign="center" paddingVertical="$spacing16">
+            Bonding curve complete! Graduate to enable trading on JuiceSwap V2.
+          </Text>
+          <GraduateButton
+            onPress={isGraduating ? undefined : handleGraduate}
+            opacity={isGraduating ? 0.6 : 1}
+            cursor={isGraduating ? 'not-allowed' : 'pointer'}
+          >
+            <Text variant="buttonLabel2" color="$white">
+              {isGraduating ? 'Graduating...' : 'Graduate to JuiceSwap V2'}
             </Text>
-          </MaxButton>
-        </InputWrapper>
-      </InputContainer>
+          </GraduateButton>
+        </>
+      ) : (
+        <>
+          <InputContainer>
+            <InputLabel>{isBuy ? 'You pay (JUSD)' : `You sell (${tokenSymbol})`}</InputLabel>
+            <InputWrapper>
+              <StyledInput
+                type="text"
+                inputMode="decimal"
+                placeholder="0.0"
+                value={amount}
+                onChange={handleAmountChange}
+                autoComplete="off"
+                autoCorrect="off"
+              />
+              <MaxButton onPress={handleMaxClick}>
+                <Text variant="buttonLabel4" color="$accent1">
+                  MAX
+                </Text>
+              </MaxButton>
+            </InputWrapper>
+          </InputContainer>
 
-      <OutputCard>
-        <OutputRow>
-          <Text variant="body2" color="$neutral2">
-            {isBuy ? `You receive (${tokenSymbol})` : 'You receive (JUSD)'}
-          </Text>
-          <Text variant="body1" color="$neutral1" fontWeight="500">
-            {buyQuoteLoading || sellQuoteLoading
-              ? '...'
-              : Number(outputAmount).toLocaleString(undefined, { maximumFractionDigits: 4 })}
-          </Text>
-        </OutputRow>
+          <OutputCard>
+            <OutputRow>
+              <Text variant="body2" color="$neutral2">
+                {isBuy ? `You receive (${tokenSymbol})` : 'You receive (JUSD)'}
+              </Text>
+              <Text variant="body1" color="$neutral1" fontWeight="500">
+                {buyQuoteLoading || sellQuoteLoading
+                  ? '...'
+                  : Number(outputAmount).toLocaleString(undefined, { maximumFractionDigits: 4 })}
+              </Text>
+            </OutputRow>
 
-        <BalanceRow>
-          <Text variant="body3" color="$neutral2">
-            Your JUSD balance
-          </Text>
-          <Text variant="body3" color="$neutral1">
-            {formattedBaseBalance}
-          </Text>
-        </BalanceRow>
+            <BalanceRow>
+              <Text variant="body3" color="$neutral2">
+                Your JUSD balance
+              </Text>
+              <Text variant="body3" color="$neutral1">
+                {formattedBaseBalance}
+              </Text>
+            </BalanceRow>
 
-        <BalanceRow>
-          <Text variant="body3" color="$neutral2">
-            Your {tokenSymbol} balance
-          </Text>
-          <Text variant="body3" color="$neutral1">
-            {formattedTokenBalance}
-          </Text>
-        </BalanceRow>
-      </OutputCard>
+            <BalanceRow>
+              <Text variant="body3" color="$neutral2">
+                Your {tokenSymbol} balance
+              </Text>
+              <Text variant="body3" color="$neutral1">
+                {formattedTokenBalance}
+              </Text>
+            </BalanceRow>
+          </OutputCard>
 
-      {error && (
-        <Text variant="body3" color="$statusCritical" textAlign="center">
-          {error}
-        </Text>
+          {error && (
+            <Text variant="body3" color="$statusCritical" textAlign="center">
+              {error}
+            </Text>
+          )}
+
+          <ActionButton
+            variant={isButtonDisabled ? 'disabled' : isWalletConnected ? (isBuy ? 'buy' : 'sell') : 'connect'}
+            onPress={isButtonDisabled ? undefined : handleButtonPress}
+          >
+            <Text variant="buttonLabel2" color="$white">
+              {buttonText}
+            </Text>
+          </ActionButton>
+
+          <Text variant="body4" color="$neutral3" textAlign="center">
+            1% of all trade volume flows to JUICE governance token holders. Slippage tolerance: 1%
+          </Text>
+        </>
       )}
-
-      <ActionButton
-        variant={isButtonDisabled ? 'disabled' : isWalletConnected ? (isBuy ? 'buy' : 'sell') : 'connect'}
-        onPress={isButtonDisabled ? undefined : handleButtonPress}
-      >
-        <Text variant="buttonLabel2" color="$white">
-          {buttonText}
-        </Text>
-      </ActionButton>
-
-      <Text variant="body4" color="$neutral3" textAlign="center">
-        1% of all trade volume flows to JUICE governance token holders. Slippage tolerance: 1%
-      </Text>
     </PanelContainer>
   )
 }
